Update add-graphics-as-text
Notes on use:
When the layer and field are being selected an error message will appear in the widget.
This appears to be a bug. Refreshing the page will remove the error.

Make sure 'selected features' is applied when choosing a data source.

Because the printed map does not always reflect the visible area in the webmap the graphic has been added either to the bottom of the page on portrait and to the side on landscape.
The positioning of the graphics can be adjusted in the code as required.

Instability in the display of the graphics has been experienced in the past so thorough testing is recommended before rollout.